I wrote out a new plan for the threads, managers and objects
This should fix most issues:

First step: Convert all structs to Obj-C classes. I don't want to have
to deal with memory management
Second step: Change the detection manager so the sub threads don't
syncronize, they just alter the detection manager variables. As long as
the variables are not used between sub threads we should be fine.
Third step: I don't think it will be necessary to syncronize the
retrieval of detection data. Because for arrays I'm just going to
replace the pointer to a new array. Which means the AI logic needs to
pull once on the data. If it pulls multiple times while doing logic, it
could pull different data.

This will allow the Detection Manager to run at full speed and the the
AI Logic to run at full speed.